backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
neo4j==5.17.0
python-dotenv==1.0.1
pydantic==2.6.1
networkx==3.2.1
torch==2.2.0
torch-geometric==2.5.0
transformers==4.37.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
community==1.0.0b1
node2vec==0.4.6
